+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-CHANGELOG
-==========
+# CHANGELOG
 
+- 2026-05-25 wsEventEmitter: Add /raw_io_board_state event emitter
+- 2026-05-25 topicMessages: Add IoBoardStateMsg interface
 - 2026-05-22 wsEventEmitter: Add /top_laser_2d/scan event emitter
 - 2026-05-22 wsEventEmitter: Add /top_laser_3d/scan event emitter
 - 2026-05-22 robotApi: Add calibrateTopLidar service API